requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
idna==3.3
Pillow==9.0.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
